+++ relay-dynamic-config/src/global.rs
@@ -180,6 +180,16 @@ pub struct Options {
     )]
     pub unreal_report_expansion_rollout_rate: f32,
 
+    /// Rollout rate for fetching the project config in the minidump endpoint.
+    ///
+    /// Rate needs to be between `0.0` and `1.0`.
+    #[serde(
+        rename = "relay.minidump-endpoint-fetch-config.rollout-rate",
+        deserialize_with = "default_on_error",
+        skip_serializing_if = "is_default"
+    )]
+    pub minidump_endpoint_fetch_config_rollout_rate: f32,
+
     /// All other unknown options.
     #[serde(flatten)]
     other: HashMap<String, Value>,